requirements_full.txt
torch>=2.0.0
sentence-transformers>=2.2.2
transformers>=4.30.0
huggingface-hub>=0.16.0
tokenizers>=0.13.0
accelerate>=0.20.0
chromadb>=0.4.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
google-generativeai
langchain>=0.0.200
langchain-experimental>=0.0.20
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
mammoth>=1.6.0
beautifulsoup4>=4.12.0
nltk>=3.8
spacy>=3.6.0
streamlit>=1.25.0
gradio>=3.35.0
jupyter>=1.0.0
ipykernel>=6.16.0
matplotlib>=3.6.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
aiofiles>=23.0.0
wikipediaapi>=0.6.0
asyncio-throttle>=1.0.0
streamlit>=1.25.0